Try with `onClose` checks

+++ lib/domain/service/auth.dart
@@ -262,6 +262,10 @@ class AuthService extends Dependency {
     });
 
     return await WebUtils.protect(() async {
+      if (isClosed) {
+        return null;
+      }
+
       final List<Credentials> allCredentials = await _credentialsProvider.all();
       for (final Credentials e in allCredentials) {
         WebUtils.putCredentials(e);
@@ -402,6 +406,10 @@ class AuthService extends Dependency {
     status.value = force ? RxStatus.loadingMore() : RxStatus.loading();
 
     await WebUtils.protect(() async {
+      if (isClosed) {
+        return;
+      }
+
       // If service is already authorized, then no-op, as this operation is
       // meant to be invoked only during unauthorized phase or account
       // switching, or otherwise the dependencies will be broken as of now.
@@ -497,6 +505,10 @@ class AuthService extends Dependency {
 
     status.value = RxStatus.loadingMore();
     await WebUtils.protect(() async {
+      if (isClosed) {
+        return;
+      }
+
       await _authorized(credentials);
       status.value = RxStatus.success();
     });
@@ -560,6 +572,10 @@ class AuthService extends Dependency {
     }
 
     return await WebUtils.protect(() async {
+      if (isClosed) {
+        return null;
+      }
+
       try {
         await _authRepository.deleteSession();
       } catch (e) {
@@ -628,6 +644,10 @@ class AuthService extends Dependency {
       final bool areValid = await validateToken(creds);
       if (areValid) {
         await WebUtils.protect(() async {
+          if (isClosed) {
+            return;
+          }
+
           await _authorized(creds!);
           status.value = RxStatus.success();
         });
@@ -674,6 +694,10 @@ class AuthService extends Dependency {
     }
 
     return await WebUtils.protect(() async {
+      if (isClosed) {
+        return false;
+      }
+
       // If [creds] are not provided, then validate the current [credentials].
       creds ??= credentials.value;
 
@@ -721,6 +745,10 @@ class AuthService extends Dependency {
       // Wait for the lock to be released and check the [Credentials] again as
       // some other task may have already refreshed them.
       await WebUtils.protect(() async {
+        if (isClosed) {
+          return;
+        }
+
         Log.debug(
           'refreshSession($userId |-> $attempt) acquired both `dbLock` and `WebUtils.protect()`',
           '$runtimeType',
